+++ backend/src/main/java/com/kindejak/javalinapp/request/BasicRequest.java
@@ -1,4 +1,4 @@
-package com.kindejak.javalinapp.utils;
+package com.kindejak.javalinapp.request;
 
 import lombok.Getter;
 import lombok.Setter;
